integration test added for endpoint health_check
