+++ src/lib.rs
@@ -1,34 +1,36 @@
-pub fn add(a: i32, b: i32) -> i32 {
-    a + b
+use actix_web::{get, post, App, HttpResponse, HttpServer, Responder};
+
+#[get("/")]
+async fn hello() -> impl Responder {
+    HttpResponse::Ok().body("Hello World.")
 }
 
-pub fn subtract(a: i32, b: i32) -> i32 {
-    a - b
+#[get("/greeting")]
+async fn greeting() -> impl Responder {
+    HttpResponse::Ok().body("Greeting!")
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_add() {
-        let result = add(2, 4);
-        assert_eq!(result, 6);
-    }
-
-    #[test]
-    fn add_negatives() {
-        let ans = add(-1, -4);
-        assert_eq!(ans, -5);
-    }
+// to test post enpoint, in terminal
+// curl -X POST 127.0.0.1:8080/name -H "Content-Type: plain/text" -d "request body here"
+#[post("/name")]
+async fn name(req_body: String) -> impl Responder {
+    HttpResponse::Ok().body(req_body)
 }
 
-#[cfg(test)]
-mod subtract {
-    use crate::subtract;
+#[get("/health_check")]
+async fn health_check() -> impl Responder {
+    HttpResponse::Ok().finish()
+}
 
-    #[test]
-    fn test_subtract() {
-        assert_eq!(7, subtract(10, 3))
-    }
+pub async fn run() -> std::io::Result<()> {
+    HttpServer::new(|| {
+        App::new()
+            .service(hello)
+            .service(greeting)
+            .service(name)
+            .service(health_check)
+    })
+    .bind(("0.0.0.0", 8080))?
+    .run()
+    .await
 }
