--- tests/integration_test.rs
@@ -1,6 +1,0 @@
-use zero2prod::add;
-
-#[test]
-fn test_add_integration() {
-    assert_eq!(add(5, 7), 12);
-}
